refactor main a routers
