+++ .github/workflows/test_back.yml
@@ -1,0 +1,28 @@
+name: Python application
+
+on: 
+  push:
+    branches: '**'
+
+jobs:
+  build:
+    runs-on: ubuntu-latest
+
+    steps:
+    - name: Check out the code
+      uses: actions/checkout@v3
+      with:
+        ref: ${{ github.ref_name }}  # Asegura usar la misma rama que hizo push
+
+    - name: Set up Python
+      uses: actions/setup-python@v4
+      with:
+        python-version: 3.10
+
+    - name: Install dependencies
+      run: |
+        python -m pip install --upgrade pip
+        pip install -r requirements.txt
+
+    - name: Run tests
+      run: pytest
