+++ src/routers/home.py
@@ -1,0 +1,102 @@
+from fastapi import FastAPI, Request, Depends, status, HTTPException, APIRouter
+from fastapi import WebSocket, WebSocketDisconnect
+from fastapi.middleware.cors import CORSMiddleware
+from fastapi.responses import JSONResponse
+from pydantic import ValidationError, BaseModel
+
+from sqlalchemy.exc import SQLAlchemyError, IntegrityError
+from src.db import Base, engine, SessionLocal, get_db
+from sqlalchemy.orm import Session
+
+from src.models.jugadores import Jugador
+from src.models.partida import Partida
+from src.models.utils import Partida_config, Leave_config
+from src.models.tablero import Tablero
+from src.models.cartafigura import PictureCard
+from src.models.cartamovimiento import MovementCard
+from src.models.fichas_cajon import FichaCajon
+from src.ws_manager import WebSocketConnectionManager
+from src.models.events import Event
+
+from src.repositories.board_repository import *
+from src.repositories.game_repository import *
+from src.repositories.player_repository import *
+from src.repositories.cards_repository import *
+from src.game import *
+
+router = APIRouter()
+ws_manager = WebSocketConnectionManager()
+event = Event()
+
+@router.websocket("")
+async def websocket_endpoint(websocket: WebSocket):
+    """
+    Le permite al front escucha los mensajes entrantes, que se envían a
+    todos aquellos en home
+    """
+    await ws_manager.connect(game_id=0, websocket=websocket)
+    try:
+        while True:
+            data = await websocket.receive_text()
+    except WebSocketDisconnect:
+        ws_manager.disconnect(websocket)
+        await ws_manager.send_all_message("Un usuario se ha desconectado")
+
+
+@router.get("/get-lobbies", status_code=status.HTTP_200_OK)
+async def get_lobbies(db: Session = Depends(get_db)):
+    """
+    Descripción: maneja la logica de pedir las lobbies al servidor.
+
+    Respuesta:
+    - 200: OK
+    - 500: Ocurre un error interno.
+    """
+    try:
+        lobbies = list_lobbies(db)
+    except Exception:
+        raise HTTPException(status_code=status.HTTP_500_INTERNAL_SERVER_ERROR, detail="Error al obtener los lobbies.")
+    return lobbies
+
+
+@router.get("/lobby", status_code=status.HTTP_200_OK)
+async def get_lobby_info(game_id: GameId = Depends(), db: Session = Depends(get_db)):
+    """
+    Descripción: maneja la logica de pedir al servidor la información de una
+    lobby (jugadores, partida, cantidad maxima, etc).
+
+    Respuesta:
+    - 200: OK.
+    - 500: Ocurre un error interno.
+    """
+    try:
+        lobby_info = get_lobby(game_id.game_id, db)    
+    except Exception:
+        raise HTTPException(status_code=status.HTTP_500_INTERNAL_SERVER_ERROR, detail="Error al obtener la partida")
+    
+    return lobby_info
+
+
+@router.post("/create-config", status_code=status.HTTP_201_CREATED)
+async def create_partida(partida_config: Partida_config, db: Session = Depends(get_db)):
+    """
+    Descripción: maneja la logica de crear una partida mediante un cierta configuración
+
+    Respuesta:
+    - 201: OK la partida fue creada.
+    - 500: Ocurre un erro interno.
+    """
+    try:
+        id_game = add_partida(partida_config, db)
+
+        #Luego de crear la partida, le actualizo a los ws conectados la nueva lista de lobbies
+        await ws_manager.send_message_game_id(event.get_lobbies, game_id = 0)
+
+    except SQLAlchemyError:
+        db.rollback()
+        raise HTTPException(status_code=status.HTTP_500_INTERNAL_SERVER_ERROR, detail="Fallo en la base de datos")
+    
+    return JSONResponse(
+        content={"id": id_game},
+        status_code=status.HTTP_201_CREATED
+    )
