+++ tests/test_agregar_jugador.py
@@ -17,12 +17,12 @@ def client():
 
 def test_join_success(client):
     # Mock de la sesión de la base de datos
-    with patch('src.main.get_db'): 
+    with patch('src.db.get_db'): 
         player_mock = MagicMock()
         player_mock.return_value = Jugador(id= 1, nombre= "testjoin")
         
-        with patch('src.main.add_player_game'):
-            with patch('src.main.get_Partida', return_value= Partida(id= 1, game_name="partida", max_players= 4)):
+        with patch('src.routers.game.add_player_game'):
+            with patch('src.routers.game.get_Partida', return_value= Partida(id= 1, game_name="partida", max_players= 4)):
                 response = client.post("/game/join", json={"player_id": player_mock.return_value.id , "game_id": 1})
   
                 assert response.status_code == 200
@@ -30,8 +30,8 @@ def test_join_success(client):
                                     "game_id": 1}
 
 def test_join_many_players(client):
-    with patch('src.main.add_player_game') as mock_jugador:
-        with patch('src.main.get_Partida', return_value= Partida(id= 1, game_name="partida", max_players= 4)):
+    with patch('src.routers.game.add_player_game') as mock_jugador:
+        with patch('src.routers.game.get_Partida', return_value= Partida(id= 1, game_name="partida", max_players= 4)):
             mock_jugador.return_value = Jugador(id= 1, nombre= "testjoins")
             mock_jugador.return_value.partida_id = 1
             response = client.post("/game/join", json={"player_id": 1 , "game_id": 1})
@@ -48,12 +48,12 @@ def test_join_many_players(client):
             assert response1.json() == {"player_id": 3, "game_id": 1}
 
 def test_join_failure(client):
-    with patch('src.main.get_db'):
+    with patch('src.db.get_db'):
         player_mock = MagicMock()
         player_mock.return_value = Jugador(id= 1, nombre= "testjoin")
 
-        with patch('src.main.add_player_game', side_effect=SQLAlchemyError()):
-            with patch('src.main.get_Partida', return_value= Partida(id= 1, game_name="partida", max_players= 4)):
+        with patch('src.routers.game.add_player_game', side_effect=SQLAlchemyError()):
+            with patch('src.routers.game.get_Partida', return_value= Partida(id= 1, game_name="partida", max_players= 4)):
         
                 response = client.post("/game/join", json={"player_id": player_mock.return_value.id , "game_id": 1})
 
@@ -61,16 +61,16 @@ def test_join_failure(client):
                 assert response.json() == {"detail": "Error al unirse a partida"}
 
 def test_player_not_found(client):
-    with patch('src.main.add_player_game', return_value= None):
-        with patch('src.main.get_Partida', return_value= Partida(id= 1, game_name="partida", max_players= 4)):
+    with patch('src.routers.game.add_player_game', return_value= None):
+        with patch('src.routers.game.get_Partida', return_value= Partida(id= 1, game_name="partida", max_players= 4)):
             response = client.post("/game/join", json={"player_id": 1 , "game_id": 1})
 
             assert response.status_code == 404
             assert response.json() == {"detail": "El jugador no existe"}
 
 def test_game_not_found(client):
-    with patch('src.main.add_player_game', return_value= Jugador(id= 1, nombre= "test")):
-        with patch('src.main.get_Partida', return_value= None):
+    with patch('src.routers.game.add_player_game', return_value= Jugador(id= 1, nombre= "test")):
+        with patch('src.routers.game.get_Partida', return_value= None):
             response = client.post("/game/join", json={"player_id": 1 , "game_id": 1})
 
             assert response.status_code == 404
