+++ tests/test_enviar_jugador_turno.py
@@ -38,8 +38,8 @@ def test_enviar_jugador_turno_succesful(test_db, client):
     test_db.add(jugador4)
     test_db.commit()
 
-    with patch("src.main.get_db"):
-        with patch("src.main.jugador_en_turno") as mock_jugador_en_turno:
+    with patch("src.db.get_db"):
+        with patch("src.routers.game.jugador_en_turno") as mock_jugador_en_turno:
             mock_jugador_en_turno.return_value = jugador_en_turno(partida.id, test_db)
 
             response = client.get("/game/current-turn?game_id=" + str(partida.id))
@@ -54,8 +54,8 @@ def test_enviar_jugador_turno_succesful(test_db, client):
 
 def test_enviar_jugador_turno_unsuccesful(client):
 
-    with patch("src.main.get_db"):
-        with patch("src.main.jugador_en_turno") as mock_jugador_en_turno:
+    with patch("src.db.get_db"):
+        with patch("src.routers.game.jugador_en_turno") as mock_jugador_en_turno:
             mock_jugador_en_turno.side_effect = IntegrityError("Error de DB", params=None, orig=None)
 
             response = client.get("/game/current-turn?game_id=1")
