+++ tests/test_enviar_tablero.py
@@ -11,7 +11,7 @@ from src.models.cartamovimiento import MovementCard
 from src.models.fichas_cajon import FichaCajon
 from src.models.color_enum import Color
 from src.main import app
-from src.main import game_manager
+from src.routers.game import game_manager
 from src.repositories.board_repository import get_fichas
 from fastapi.testclient import TestClient
 from unittest.mock import patch
@@ -43,8 +43,8 @@ def test_enviar_tablero_succesful_real(test_db, client):
             test_db.refresh(ficha)
 
 
-    with patch("src.main.get_db"):
-        with patch("src.main.get_fichas") as mock_get_fichas:
+    with patch("src.db.get_db"):
+        with patch("src.routers.game.get_fichas") as mock_get_fichas:
             mock_get_fichas.return_value = get_fichas(partida.id, test_db)
 
             response = client.get("/game/board?game_id=" + str(partida.id))
@@ -63,8 +63,8 @@ def test_enviar_tablero_succesful_real(test_db, client):
 
 def test_enviar_tablero_unsuccesful(client):
 
-    with patch("src.main.get_db"):
-        with patch("src.main.get_fichas") as mock_terminar_turno:
+    with patch("src.db.get_db"):
+        with patch("src.routers.game.get_fichas") as mock_terminar_turno:
             mock_terminar_turno.side_effect = IntegrityError("Error de DB", params=None, orig=None)
 
             response = client.get("/game/board?game_id=1")
