+++ tests/test_get_lobbies.py
@@ -9,24 +9,19 @@ def client():
     return TestClient(app)
 
 def test_get_lobbies_succes(client):
-    with patch('src.main.get_db'):
-#I want to mock the database too:
-        with patch('src.main.list_lobbies', return_value=[]):
+    with patch('src.db.get_db'):
+        with patch('src.routers.home.list_lobbies', return_value=[]):
             response = client.get("/home/get-lobbies")
 
             assert response.status_code == 200
             assert response.json() == []
 
 def test_get_lobbies_failure(client):
-    with patch('src.main.get_db'):
-
-#I want to mock an exception in the list_lobbies function:
-        with patch('src.main.list_lobbies', side_effect=OperationalError("Error de DB", 
+    with patch('src.db.get_db'):
+        with patch('src.routers.home.list_lobbies', side_effect=OperationalError("Error de DB", 
                                                                         params=None, 
                                                                         orig=None)):
         
-
-
             response = client.get("/home/get-lobbies")
 
             assert response.status_code == 500
