+++ tests/test_terminar_turno.py
@@ -9,7 +9,7 @@ from src.models.cartafigura import PictureCard
 from src.models.tablero import Tablero
 from src.models.cartamovimiento import MovementCard
 from src.models.fichas_cajon import FichaCajon
-from src.main import app, end_turn 
+from src.main import app
 from src.repositories.board_repository import mezclar_fichas
 from src.repositories.game_repository import terminar_turno
 from fastapi.testclient import TestClient
@@ -42,11 +42,11 @@ def test_terminar_turno_succesful(test_db, client):
     #next_player = test_db.query(Jugador).filter(Jugador.partida_id == partida.id, Jugador.turno == id_next_player).first()
 
     #Testeo el endpoint usando la base de datos que cree
-    with patch("src.main.get_db"), \
-         patch("src.main.get_current_turn_player"), \
-         patch("src.main.game_manager.is_tablero_parcial", return_value=False), \
-         patch("src.main.terminar_turno") as mock_terminar_turno, \
-         patch("src.main.repartir_cartas", return_value= None):
+    with patch("src.db.get_db"), \
+         patch("src.routers.game.get_current_turn_player"), \
+         patch("src.routers.game.game_manager.is_tablero_parcial", return_value=False), \
+         patch("src.routers.game.terminar_turno") as mock_terminar_turno, \
+         patch("src.routers.game.repartir_cartas", return_value= None):
         
         mock_terminar_turno.return_value = terminar_turno(partida.id, test_db)
 
@@ -58,8 +58,8 @@ def test_terminar_turno_succesful(test_db, client):
 
 def test_terminar_turno_unsuccesful(test_db, client):
 
-    with patch("src.main.get_db"):
-        with patch("src.main.terminar_turno") as mock_terminar_turno:
+    with patch("src.db.get_db"):
+        with patch("src.routers.game.terminar_turno") as mock_terminar_turno:
             mock_terminar_turno.side_effect = IntegrityError("Error de DB", params=None, orig=None)
 
             response = client.put("/game/end-turn", json={"game_id": 1})
